Implemented AbstractRavMoves
